Section 2, Lesson 13 Boolean logic and Switch cases
Also, messed around with functions and while statements to see if it works similar to python.  Big finding is difference in how global and local variables are handeled.  While in many or most other programming languages variables are treated as global if not otherwise declared, Python deals with variables the other way around. They are local, if not otherwise declared.